editoast: move track_offset to infra editoast_schemas
