+++ editoast/editoast_schemas/src/infra.rs
@@ -1,0 +1,7 @@
+mod track_offset;
+
+pub use track_offset::TrackOffset;
+
+editoast_common::schemas! {
+    track_offset::schemas(),
+}
