+++ editoast/editoast_schemas/src/lib.rs
@@ -1,11 +1,11 @@
+pub mod infra;
 pub mod primitives;
 pub mod rolling_stock;
-pub mod track_offset;
 pub mod train_schedule;
 
 editoast_common::schemas! {
     rolling_stock::schemas(),
     train_schedule::schemas(),
-    track_offset::schemas(),
     primitives::schemas(),
+    infra::schemas(),
 }
