+++ editoast/src/views/v2/path/pathfinding.rs
@@ -9,8 +9,8 @@ use actix_web::web::Json;
 use actix_web::web::Path;
 use diesel_async::AsyncPgConnection as PgConnection;
 use editoast_common::Identifier;
+use editoast_schemas::infra::TrackOffset;
 use editoast_schemas::rolling_stock::LoadingGaugeType;
-use editoast_schemas::track_offset::TrackOffset;
 use editoast_schemas::train_schedule::PathItemLocation;
 use serde::Deserialize;
 use serde::Serialize;
